AmiCo and levels update

+++ trunk/src/ch/idsia/ai/agents/ai/SimpleMLPAgent.java
@@ -34,7 +34,7 @@ public class  SimpleMLPAgent implements Agent, Evolvable {
         return new SimpleMLPAgent (mlp.copy ());
     }
 
-    public void integrateObservation(byte[] serializedLevelSceneObservationZ, byte[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
+    public void integrateObservation(int[] serializedLevelSceneObservationZ, int[] serializedEnemiesObservationZ, float[] marioFloatPos, float[] enemiesFloatPos, int[] marioState)
     {
         //To change body of implemented methods use File | Settings | File Templates.
     }
